% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryxclass_adj.R
\name{adjustment}
\alias{adjustment}
\alias{Tryx}
\title{R6 Class for outlier adjustment estimation}
\description{
How much of the heterogeneity due to the outlier can be explained by alternative pathways?
}
\section{Usage}{
 

x$adjustment(dat= self$output$dat, tryxscan=self$output, id_remove=NULL)
}

\section{Arguments}{

\code{tryxscan} Output from \code{Tryx}

\code{id_remove} List of IDs to exclude from the adjustment analysis. It is possible that in the outlier search a candidate trait will come up which is essentially just a surrogate for the outcome trait (e.g. if you are analysing coronary heart disease as the outcome then a variable related to heart disease medication might come up as a candidate trait). Adjusting for a trait which is essentially the same as the outcome will erroneously nullify the result, so visually inspect the candidate trait list and remove those that are inappropriate.

\code{proxies} Look for proxies in the MVMR methods. Default = FALSE.
}

