% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryxclass_adj.R
\name{analyse}
\alias{analyse}
\title{Analyse tryx results}
\description{
This returns various heterogeneity statistics, IVW estimates for raw, 
adjusted and outlier removed datasets, and summary of peripheral 
traits detected etc.
}
\section{Usage}{

}

