% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.r
\name{cochrans_q}
\alias{cochrans_q}
\title{Cochran's Q statistic}
\usage{
cochrans_q(b, se)
}
\arguments{
\item{b}{vector of effecti}

\item{se}{vector of standard errors}
}
\value{
q values
}
\description{
Cochran's Q statistic
}
