% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.r
\name{simulate.tryx}
\alias{simulate.tryx}
\title{Simulate data to test tryx}
\usage{
\method{simulate}{tryx}(nid, nu1, nu2, bxy = 3, outliers_known = TRUE)
}
\arguments{
\item{nid}{Sample size to simulate}

\item{nu1}{Number of traits which influence both X and Y}

\item{nu2}{Number of traits which influence Y}

\item{bxy}{Causal effect of X on Y}

\item{outliers_known}{Assume that all outliers are detected (defaukt = TRUE). If FALSE then Radial MR is used to detect outliers}
}
\value{
Results from simulations and tryx scan
}
\description{
Choose a sample size, number of horizontal pleiotropy paths, and number of confounders. Create summary data for all necessary traits, then perform tryx scan
}
