% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{tryx.network}
\alias{tryx.network}
\title{Plot results from outlier_scan in a network}
\usage{
tryx.network(tryxscan)
}
\arguments{
\item{tryxscan}{Output from outlier_scan function}
}
\value{
Prints plot, and returns dataframe of the connections
}
\description{
Creates a simple network depicting the connections between outlier instruments, the original exposure and outcome traits, and the detected candidate associations
}
