% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryx-package.R
\docType{package}
\name{tryx-package}
\alias{tryx}
\alias{tryx-package}
\title{tryx: MR-TRYX (treasure your exceptions)}
\description{
Heterogeneity in MR analyses can arise due to horizontal pleiotropy. This package uses MR-Base to identify possible traits that can explain the heterogeneity, with a view to identifying novel putative associations, and adjusting for their influences to reduce heterogeneity and improve power.
}
\author{
\strong{Maintainer}: Gibran Hemani \email{g.hemani@bristol.ac.uk}

Authors:
\itemize{
  \item Gibran Hemani \email{g.hemani@bristol.ac.uk}
  \item Yoonsu Cho \email{yoonsu.cho@bristol.ac.uk}
}

}
\keyword{internal}
