% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.r
\name{tryx.adjustment}
\alias{tryx.adjustment}
\title{Outlier adjustment estimation}
\usage{
tryx.adjustment(tryxscan, id_remove = NULL)
}
\arguments{
\item{tryxscan}{Output from \code{tryx.scan}}
}
\value{
data frame of adjusted effect estimates and heterogeneity stats
}
\description{
How much of the heterogeneity due to the outlier can be explained by alternative pathways?
}
