% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.r
\name{tryx.analyse}
\alias{tryx.analyse}
\title{Analyse tryx results}
\usage{
tryx.analyse(tryxscan, plot = TRUE, id_remove = NULL,
  filter_duplicate_outliers = TRUE)
}
\arguments{
\item{tryxscan}{Output from \code{tryx.scan}}

\item{plot}{Whether to plot or not. Default is TRUE}

\item{id_remove}{List of IDs to exclude from the adjustment analysis. It is possible that in the outlier search a candidate trait will come up which is essentially just a surrogate for the outcome trait (e.g. if you are analysing coronary heart disease as the outcome then a variable related to heart disease medication might come up as a candidate trait). Adjusting for a trait which is essentially the same as the outcome will erroneously nullify the result, so visually inspect the candidate trait list and remove those that are inappropriate.}

\item{duplicate_outliers_method}{Sometimes more than one trait will associate with a particular outlier. TRUE = only keep the trait that has the biggest influence on heterogeneity}
}
\value{
List of 
- adj_full: data frame of SNP adjustments for all candidate traits
- adj: The results from adj_full selected to adjust the exposure-outcome model
- Q: Heterogeneity stats
- estimates: Adjusted and unadjested exposure-outcome effects
- plot: Radial plot showing the comparison of different methods and the changes in SNP effects ater adjustment
}
\description{
This returns various heterogeneity statistics, IVW estimates for raw, 
adjusted and outlier removed datasets, and summary of peripheral 
traits detected etc.
}
