% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.r
\name{tryx.analyse}
\alias{tryx.analyse}
\title{Analyse tryx results}
\usage{
tryx.analyse(tryxscan, plot = TRUE, filter_duplicate_outliers = TRUE)
}
\arguments{
\item{tryxscan}{Output from \code{tryx.scan}}

\item{plot}{Whether to plot or not. Default is TRUE}

\item{filter_duplicate_outliers}{Whether to only allow each putative outlier to be adjusted by a single trait (in order of largest divergence). Default is TRUE.}
}
\description{
This returns various heterogeneity statistics, IVW estimates for raw, 
adjusted and outlier removed datasets, and summary of peripheral 
traits detected etc.
}
