% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryxclass.R
\docType{data}
\name{Tryx}
\alias{Tryx}
\title{Class for MR-TRYX analysis}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tryx
}
\arguments{
\item{dat}{Dataset from TwoSampleMR::harmonise_data}

\item{description}{Detect outliers in exposure-outcome dataset}

\item{outliers}{Default is to use the RadialMR package to identify IVW outliers. Alternatively can providen an array of SNP names that are present in dat$SNP to use as outliers.}

\item{outlier_correction}{Defualt = "none", but can select from ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none").}

\item{outlier_threshold}{If outlier_correction = "none" then the p-value threshold for detecting outliers is by default 0.05.}
}
\description{
Create a new dataset
}
\details{
A simple wrapper function.
Using a summary set, find outliers in the MR analysis between the pair of trais.
Find other 'candidate traits' associated with those outliers.
Perform MR of each of those candidate traits with the original exposure and outcome.
}
\keyword{datasets}
