% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryxclass.R
\name{Tryx}
\alias{Tryx}
\title{Class for MR-TRYX analysis}
\description{
A simple wrapper function.
Using a summary set, find outliers in the MR analysis between the pair of trais.
Find other 'candidate traits' associated with those outliers.
Perform MR of each of those candidate traits with the original exposure and outcome.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Tryx$new()}}
\item \href{#method-print}{\code{Tryx$print()}}
\item \href{#method-get_outliers}{\code{Tryx$get_outliers()}}
\item \href{#method-set_candidate_traits}{\code{Tryx$set_candidate_traits()}}
\item \href{#method-scan}{\code{Tryx$scan()}}
\item \href{#method-candidate_instruments}{\code{Tryx$candidate_instruments()}}
\item \href{#method-outcome_instruments}{\code{Tryx$outcome_instruments()}}
\item \href{#method-exposure_instruments}{\code{Tryx$exposure_instruments()}}
\item \href{#method-exposure_candidate_instruments}{\code{Tryx$exposure_candidate_instruments()}}
\item \href{#method-extractions}{\code{Tryx$extractions()}}
\item \href{#method-candidate_outcome_dat}{\code{Tryx$candidate_outcome_dat()}}
\item \href{#method-candidate_exposure_dat}{\code{Tryx$candidate_exposure_dat()}}
\item \href{#method-exposure_candidate_dat}{\code{Tryx$exposure_candidate_dat()}}
\item \href{#method-harmonise}{\code{Tryx$harmonise()}}
\item \href{#method-mr}{\code{Tryx$mr()}}
\item \href{#method-mrtryx}{\code{Tryx$mrtryx()}}
\item \href{#method-tryx.sig}{\code{Tryx$tryx.sig()}}
\item \href{#method-adjustment}{\code{Tryx$adjustment()}}
\item \href{#method-adjustment.mv}{\code{Tryx$adjustment.mv()}}
\item \href{#method-analyse}{\code{Tryx$analyse()}}
\item \href{#method-analyse.mv}{\code{Tryx$analyse.mv()}}
\item \href{#method-manhattan_plot}{\code{Tryx$manhattan_plot()}}
\item \href{#method-clone}{\code{Tryx$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new dataset and initialise an R interface
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$new(dat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Dataset from TwoSampleMR::harmonise_data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_outliers"></a>}}
\subsection{Method \code{get_outliers()}}{
Detect outliers in exposure-outcome dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$get_outliers(
  dat = self$output$dat,
  outliers = "RadialMR",
  outlier_correction = "none",
  outlier_threshold = ifelse(outlier_correction == "none", 0.05/nrow(dat), 0.05)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Output from TwoSampleMR::harmonise_data. Note - only the first id.exposure - id.outcome pair will be used.}

\item{\code{outliers}}{Default is to use the RadialMR package to identify IVW outliers. Alternatively can providen an array of SNP names that are present in dat$SNP to use as outliers.}

\item{\code{outlier_correction}}{Defualt = "none", but can select from ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none").}

\item{\code{outlier_threshold}}{If outlier_correction = "none" then the p-value threshold for detecting outliers is by default 0.05.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_candidate_traits"></a>}}
\subsection{Method \code{set_candidate_traits()}}{
Set a list of GWAS IDs used
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$set_candidate_traits(id_list = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_list}}{The list of trait IDs to search through for candidate associations. The default is the high priority traits in available_outcomes().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scan"></a>}}
\subsection{Method \code{scan()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$scan(
  dat = self$output$dat,
  search_correction = "none",
  search_threshold = ifelse(search_correction == "none", 5e-08, 0.05),
  use_proxies = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-candidate_instruments"></a>}}
\subsection{Method \code{candidate_instruments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$candidate_instruments(
  candidate_instruments = NULL,
  include_outliers = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-outcome_instruments"></a>}}
\subsection{Method \code{outcome_instruments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$outcome_instruments(
  candidate_outcome = NULL,
  dat = self$output$dat,
  use_proxies = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exposure_instruments"></a>}}
\subsection{Method \code{exposure_instruments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$exposure_instruments(
  candidate_exposure = NULL,
  dat = self$output$dat,
  use_proxies = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exposure_candidate_instruments"></a>}}
\subsection{Method \code{exposure_candidate_instruments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$exposure_candidate_instruments(
  exposure_candidate = NULL,
  dat = self$output$dat,
  use_proxies = FALSE,
  include_outliers = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extractions"></a>}}
\subsection{Method \code{extractions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$extractions(
  dat = self$output$dat,
  candidate_instruments = NULL,
  candidate_outcome = NULL,
  candidate_exposure = NULL,
  exposure_candidate = NULL,
  include_outliers = FALSE,
  use_proxies = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-candidate_outcome_dat"></a>}}
\subsection{Method \code{candidate_outcome_dat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$candidate_outcome_dat(dat = self$output$dat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-candidate_exposure_dat"></a>}}
\subsection{Method \code{candidate_exposure_dat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$candidate_exposure_dat(dat = self$output$dat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exposure_candidate_dat"></a>}}
\subsection{Method \code{exposure_candidate_dat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$exposure_candidate_dat(dat = self$output$dat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-harmonise"></a>}}
\subsection{Method \code{harmonise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$harmonise(dat = self$output$dat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mr"></a>}}
\subsection{Method \code{mr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$mr(dat = self$output$dat, mr_method = "mr_ivw")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mrtryx"></a>}}
\subsection{Method \code{mrtryx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$mrtryx(
  dat = self$output$dat,
  outliers = "RadialMR",
  outlier_correction = "none",
  outlier_threshold = ifelse(outlier_correction == "none", 0.05/nrow(dat), 0.05),
  use_proxies = FALSE,
  search_correction = "none",
  search_threshold = ifelse(search_correction == "none", 5e-08, 0.05),
  include_outliers = FALSE,
  mr_method = "mr_ivw"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tryx.sig"></a>}}
\subsection{Method \code{tryx.sig()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$tryx.sig(mr_threshold_method = "fdr", mr_threshold = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-adjustment"></a>}}
\subsection{Method \code{adjustment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$adjustment(
  dat = self$output$dat,
  tryxscan = self$output,
  id_remove = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-adjustment.mv"></a>}}
\subsection{Method \code{adjustment.mv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$adjustment.mv(
  dat = self$output$dat,
  tryxscan = self$output,
  lasso = TRUE,
  id_remove = NULL,
  proxies = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-analyse"></a>}}
\subsection{Method \code{analyse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$analyse(
  tryxscan = self$output,
  plot = TRUE,
  id_remove = NULL,
  filter_duplicate_outliers = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-analyse.mv"></a>}}
\subsection{Method \code{analyse.mv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$analyse.mv(
  tryxscan = self$output,
  lasso = TRUE,
  plot = TRUE,
  id_remove = NULL,
  proxies = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-manhattan_plot"></a>}}
\subsection{Method \code{manhattan_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$manhattan_plot(
  dat = self$output$candidate_outcome_mr,
  id_remove = NULL,
  y_scale = NULL,
  lable = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tryx$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
