% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryxclass.R
\docType{data}
\name{Tryx}
\alias{Tryx}
\title{Class for MR-TRYX analysis}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tryx
}
\arguments{
\item{dat}{output from harmonise_data. Only allowed one exposure / outcome combination}
}
\value{
R6 class
Initialise the class with a dataset
}
\description{
<full description>
}
\keyword{datasets}
