% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.r
\name{tryx.simulate}
\alias{tryx.simulate}
\title{Simulate data to test tryx}
\usage{
tryx.simulate(nid, nu1, nu2, bxu3 = 0, bu3y = 0, bxy = 3,
  outliers_known = c("known", "detected", "all")[1], debug = FALSE)
}
\arguments{
\item{nid}{Sample size to simulate}

\item{nu1}{Number of traits which influence both X and Y}

\item{nu2}{Number of traits which influence Y}

\item{bxu3}{Effect of x on mediator}

\item{bu3y}{Effect of mediator on y}

\item{bxy}{Causal effect of X on Y}

\item{outliers_known}{Assume that all outliers are detected (defaukt = TRUE). If FALSE then Radial MR is used to detect outliers}
}
\value{
Results from simulations and tryx scan
}
\description{
Choose a sample size, number of horizontal pleiotropy paths, and number of confounders. Create summary data for all necessary traits, then perform tryx scan
}
