% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.r
\name{tryx.simulate}
\alias{tryx.simulate}
\title{Simulate data to test tryx}
\usage{
tryx.simulate(nid = 10000, ngx = 30, ngu1 = 30, ngu2 = 30,
  nu2 = 2, ngu3 = 30, vgx = 0.2, vgu1 = 0.6, vgu2 = 0.2,
  vgu3 = 0.2, bxy = 0, bu1x = 0.6, bu1y = 0.4, bxu3 = 0.3,
  bu3y = 0, vgxu2 = 0.2, vu2y = 0.2, ngxu3 = 0, vgxu3 = 0,
  mininum_instruments = 10, instrument_threshold = "bonferroni",
  outlier_threshold = "bonferroni", outliers_known = "detected",
  directional_bias = FALSE)
}
\arguments{
\item{nid}{= 10000 Number of samples}

\item{ngx}{= 30 Number of direct instruments to x}

\item{ngu1}{= 30 Number of instruments influencing confounder of x and y}

\item{ngu2}{= 30 Number of instruments per mediating}

\item{nu2}{= 2 Number of gx that are pleiotropic}

\item{ngu3}{= 30 Number of instruments for u3 mediator}

\item{vgx}{= 0.2 Variance explained by gx instruments}

\item{vgu1}{= 0.6 Variance explained by u1 instruments}

\item{vgu2}{= 0.2 Variance explained by u2 instruments}

\item{vgu3}{= 0.2 Variance explained by u3 instruments}

\item{bxy}{= 0 Causal effect of x on y}

\item{bu1x}{= 0.6 Effect of u1 on x}

\item{bu1y}{= 0.4 Effect of u1 on y}

\item{bxu3}{= 0.3 Effect of x on u3}

\item{bu3y}{= 0 Effect of u3 on y}

\item{vgxu2}{= 0.2 Variance explained by each gx instrument on each u2 mediator}

\item{vu2y}{= 0.2 Variance explained by all u2 mediators on y}

\item{ngxu3}{= 0 Number of gx instruments that pleiotropically associate with u3 mediator}

\item{mininum_instruments}{= 10 Minimum number of instruments required to have been detected to run simulation}

\item{instrument_threshold}{= "bonferroni" Threshold, either numeric or 'bonferroni'}

\item{outlier_threshold}{= "bonferroni" Threshold, either numeric or 'bonferroni'}

\item{outliers_known}{= "detected" Either detected = using radial mr to find heterogeneity outliers; known = adjust for all known invalid instruments; all = adjust for all instruments}

\item{directional_bias}{= FALSE Is the pleiotropic effect randomly centered around 0 (FALSE) or does it have a non-0 mean (TRUE)}

\item{vgxu3y}{= 0 Variance explained by all gx variants directly on u3 mediator}
}
\value{
list for tryx.analyse
}
\description{
# Create data using model here: https://www.draw.io/#G1VqjTd4iH2de7sXI4PhUrxazBnYLduH_w. Create summary data for all necessary traits, then perform tryx scan
x = exposure
y = outcome
u1 = confounder of x and y
u2 = mediator of horizontal pleiotropy from gx to y. One for each gx
u3 = mediator between x and y
}
